backend/api-gateway/requirements.txt
Django
djangorestframework
requests
gunicorn
django-cors-headers